requirements.txt
streamlit==1.19.0
altair==4.2.2
pandas
numpy
scikit-learn
matplotlib
seaborn
joblib
pillow
